weather mock data 를 viewModel 로 전달

+++ app/src/main/java/com/jin/jjinweather/layer/domain/model/weather/Weather.kt
@@ -11,5 +11,5 @@ data class Weather(
     val dailyWeatherList: List<DailyWeather>,
     val sunrise: Long,
     val sunset: Long,
-    val moonPhase: Int
+    val moonPhase: Double
 )
